requirements.txt
# Hackathon Copilot - Dependencies

# Core
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# API Client
httpx>=0.25.0

# YAML
pyyaml>=6.0.1

# Logging
loguru>=0.7.2

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend
streamlit>=1.29.0
requests>=2.31.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0